Rename files to better reflect functionality
